feat: serve actionhero instance from the process cwd
By using this type of require when we are developing plugin in actionhero we will get the live actionhero instance.
Now we can use npm link to link the plugin to a top level actionhero project and after that we can develop and test at the same time without moving a bunch of files from one location to another.
This solves the major issue that developers were facing while developing actionhero's plugins.